chore: remove mod file
